Rename Toast folder to ToastProvider. Delete Toast file. Update README.
