--- react/Toast.js
@@ -1,3 +1,0 @@
-import Toast from './components/Toast/index'
-
-export default Toast
